Add files via upload
Adding more parameters for the user data and fixing the code to work with them - read from a file, put in hashtable, etc. Also making a function which puts the final results in a file - puts the whole hashtable in a file. At the same time everything is written in the temp file at the momentg when it is done. Also i am returning the logged user in  the main function, in order to use it in future operations when the code is done.