Add files via upload
Making the menu with ANSI colours. Also fixing some things so the code looks better. Also fixing this - when the user is typing the info, the info will be added in a file after everything is typed. Fixing the name of the structures and functions. Also making parameters const when needed. Also when i had identical functions i just made them one function that just uses a parameter of length. And also mixing the salt with the password.